Feature/353 support replicating contexts (#354)

+++ packages/client-ts/src/context/engines/verida/database/db-encrypted.ts
@@ -115,23 +115,25 @@ class EncryptedDatabase extends BaseDb {
      * If there is data in this database, it ensures the current encryption key
      * can decrypt the data.
      */
-    try {
-      await this.getMany();
-      //console.log(`Db.init-5(${databaseName}): ${(new Date()).getTime()-now}`)
-    } catch (err: any) {
-      // This error message is thrown by the underlying decrypt library if the
-      // data can't be decrypted
-      if (
-        err.message == `Unsupported state or unable to authenticate data` ||
-        err.message == "Could not decrypt!"
-      ) {
-        // Clear the instantiated PouchDb instances and throw a more useful exception
-        await this.close()
-        throw new Error(`Invalid encryption key supplied`);
-      }
+    if (this.config.verifyEncryptionKey) {
+      try {
+        await this.getMany();
+        //console.log(`Db.init-5(${databaseName}): ${(new Date()).getTime()-now}`)
+      } catch (err: any) {
+        // This error message is thrown by the underlying decrypt library if the
+        // data can't be decrypted
+        if (
+          err.message == `Unsupported state or unable to authenticate data` ||
+          err.message == "Could not decrypt!"
+        ) {
+          // Clear the instantiated PouchDb instances and throw a more useful exception
+          await this.close()
+          throw new Error(`Invalid encryption key supplied`);
+        }
 
-      // Unknown error, rethrow
-      throw err;
+        // Unknown error, rethrow
+        throw err;
+      }
     }
 
     //console.log(`Db.init-Final(${databaseName}): ${(new Date()).getTime()-now}`)
@@ -228,6 +230,14 @@ class EncryptedDatabase extends BaseDb {
   public async close(options: DatabaseCloseOptions = {
     clearLocal: false
   }) {
+    if (this._sync === null) {
+      // No sync object indicates this database is closed
+      return
+    }
+
+    this._sync = null;
+    this._syncError = null;
+
     if (options.clearLocal) {
       await this.destroy({
         localOnly: true
@@ -252,8 +262,6 @@ class EncryptedDatabase extends BaseDb {
       // may already be closed
     }
 
-    this._sync = null;
-    this._syncError = null;
     await this.engine.closeDatabase(this.did, this.databaseName)
     this.emit('closed', this.databaseName)
   }
@@ -290,14 +298,14 @@ class EncryptedDatabase extends BaseDb {
     localOnly: false
   }): Promise<void> {
     try {
-      // Destory the local pouch database (this deletes this._local and this._localDbEncrypted as they share the same underlying data source)
+      // Destroy the local pouch database (this deletes this._local and this._localDbEncrypted as they share the same underlying data source)
       await this._localDbEncrypted.destroy()
 
       if (!options.localOnly) {
         // Only delete remote database if required
         await this.engine.deleteDatabase(this.databaseName)
       }
-     
+
       await this.close({
         clearLocal: false
       })
